+++ packages/staark/dst/staark.base.iife.js
@@ -167,7 +167,7 @@
             const type = typeof value;
             if (type === "function") {
               const oldValue = oldAttributes == null ? void 0 : oldAttributes[name];
-              if ((oldValue == null ? void 0 : oldValue.f) !== value.f) {
+              if ((oldValue == null ? void 0 : oldValue.f) !== value) {
                 if (oldValue) {
                   element.removeEventListener(
                     name,